Server logic for the Shiny application

This function defines the server-side logic of the Shiny application. It handles user inputs, processes data, and updates outputs accordingly. The server function sets up reactive expressions, observers, and render functions to interact with the UI components.

app_server(input, output, session)

Arguments

input

Shiny input object containing input values from the UI.

output

Shiny output object used to send output to the UI.

session

Shiny session object.

Value

None. This function is called for its side effects of setting up the server logic of the Shiny app.

Details

The server function performs several key operations:

Package Validation

Validates the selected package name and checks if it is a valid CRAN package.

Version Handling

Retrieves package versions and updates version selection inputs in the UI.

NEWS Diff Tab

Fetches the NEWS file for the package and displays the differences between versions.

Downloads Tab

Plots the download statistics of the package over time.

Checkpage Tab

Displays CRAN check results and package status.

Release List Tab

Shows the list of package releases with details.

Dependencies Tab

Displays user and developer dependencies of the package.

Info Tab

Provides package information such as title, description, version, license, etc.